Keep metadata validation runtime-free
